requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asyncio==3.4.3
CacheControl==0.14.2
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cryptography==44.0.1
cycler==0.12.1
fastapi==0.115.8
filelock==3.17.0
filetype==1.2.0
firebase-admin==6.6.0
fonttools==4.56.0
fsspec==2025.2.0
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.20.0
google-cloud-storage==3.0.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.67.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
httplib2==0.22.0
idna==3.7
Jinja2==3.1.5
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.0
mpmath==1.3.0
msgpack==1.1.0
networkx==3.4.2
numpy==2.2.2
opencv-python==4.11.0.86
opencv-python-headless==4.10.0.84
packaging==24.2
paho-mqtt==2.1.0
pillow==11.1.0
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
roboflow==1.1.53
rsa==4.9
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
sympy==1.13.1
torch==2.6.0
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
websockets==14.2
